some quick code cleanup
